Popout RoomViewer and QuickChat framework
