+++ client/src/Containers/Monitoring/QuickChat.js
@@ -1,0 +1,3 @@
+const QuickChat = () => null;
+
+export default QuickChat;
